12.6 implementation of standard layout
implemented beforeEach and describe function like every other spec files. describe only shows the name of the test in terminal for a better understanding and beforeEach makes sure the code block runs everytime an iteration happes. to implement beforeEach we created a global service variable to use it inside it function as the it function is in a different scope