12.15 writing up some not-so-effective test cases
in nest we want our controllers to be as simple as possible, we do not want logics to play on our controllers. so by nature there's not so much effective test cases to do on controller files. although we can if we want to as it is not the worst thing to do with our time